requirements.txt
pytest
python-box[all]~=6.0
requests
bs4
tqdm